# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beautiful Island Tours API Documentation",
        "description": "API documentation for Beautiful Island Tours system",
        "contact": {
            "name": "Beautiful Island Tours Support",
            "email": "ashanhimantha321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://testapi.ashanhimantha.com",
            "description": "Beautiful Island Tours API Server"
        }
    ],
    "paths": {
...
        "/api/tours": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Get list of tours",
                "description": "Display a listing of tours.",
                "operationId": "318f54f987fd4f7d2580af48cadea96c",
                "parameters": [
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter tours by location",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter tours by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "unavailable",
                                "booked",
                                "seasonal"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Tour"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tours"
                ],
                "summary": "Create new tour",
                "description": "Store a newly created tour in storage.",
                "operationId": "36df17379e2323a9bff99f881e571b1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "itinerary",
                                    "include",
                                    "exclude",
                                    "per_adult_price",
                                    "location",
                                    "status",
                                    "display_image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Island Discovery Tour"
                                    },
                                    "itinerary": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Day 1: Beach tour",
                                            "Day 2: Mountain hiking"
                                        ]
                                    },
                                    "include": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Hotel pickup",
                                            "Lunch",
                                            "Guide"
                                        ]
                                    },
                                    "exclude": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Drinks",
                                            "Tips"
                                        ]
                                    },
                                    "per_adult_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "East Coast"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "unavailable",
                                            "booked",
                                            "seasonal"
                                        ],
                                        "example": "available"
                                    },
                                    "display_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image1": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image2": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tour created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tour created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Tour"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tours/{id}/status": {
            "put": {
                "tags": [
                    "Tours"
                ],
                "summary": "Update tour status",
                "description": "Update the status of a tour.",
                "operationId": "b6f491fd8ea1dedd3be517c7e4e77e00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tour ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "unavailable",
                                            "booked",
                                            "seasonal"
                                        ],
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tour status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Tour"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Tour not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tour": {
                "required": [
                    "title",
                    "itinerary",
                    "include",
                    "exclude",
                    "per_adult_price",
                    "location"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "itinerary": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "include": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "exclude": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "per_adult_price": {
                        "type": "number",
                        "format": "float"
                    },
                    "location": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "unavailable",
                            "booked",
                            "seasonal"
                        ]
                    },
                    "display_image": {
                        "type": "string"
                    },
                    "image1": {
                        "type": "string",
                        "nullable": true
                    },
                    "image2": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}